src/release_deployment.py
```python
import os
from git import Repo

def versioning(version):
    """
    Function to handle versioning of the application.
    """
    with open("VERSION", "w") as version_file:
        version_file.write(str(version))

def testing():
    """
    Function to run tests for the application.
    """
    os.system("pytest tests/")

def ci_cd_pipeline():
    """
    Function to handle Continuous Integration and Continuous Deployment.
    """
    repo = Repo(os.getcwd())
    origin = repo.remote(name='origin')
    origin.pull()
    testing()
    origin.push()

def release(version):
    """
    Function to handle the release process.
    """
    versioning(version)
    ci_cd_pipeline()

def deployment():
    """
    Function to handle the deployment process.
    """
    os.system("docker-compose up --build -d")
```
