tests/test_release_deployment.py
```python
import unittest
from src import release_deployment

class TestReleaseDeployment(unittest.TestCase):

    def setUp(self):
        self.release = release_deployment.ReleaseDeployment()

    def test_versioning(self):
        version = self.release.get_version()
        self.assertIsInstance(version, str)

    def test_testing(self):
        test_result = self.release.run_tests()
        self.assertIsInstance(test_result, bool)
        self.assertTrue(test_result)

    def test_ci_cd_pipeline(self):
        pipeline_status = self.release.ci_cd_pipeline()
        self.assertIsInstance(pipeline_status, bool)
        self.assertTrue(pipeline_status)

if __name__ == '__main__':
    unittest.main()
```